requirements.txt
cassandra-driver
flask
